backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
python-dotenv==1.0.1
Pillow==11.1.0
pdf2image==1.17.0
pytesseract==0.3.13
google-generativeai==0.8.3
aiofiles==24.1.0
